pyproject.toml
[project]
name = "passwordsystem"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ivnstd", email = "nasty-iv@bk.ru" }
]
dependencies = [
    "uvicorn>=0.30.6",
    "fastapi>=0.115.0",
    "sqlalchemy>=2.0.35",
    "bcrypt>=4.2.0",
    "python-jose>=3.3.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.9",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
pwdsys = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/passwordsystem"]
